+++ R/theme.r
@@ -192,7 +192,7 @@ print.theme <- function(x, ...) utils::str(x)
 #'   plot.subtitle    \tab plot subtitle (text appearance)
 #'                    (\code{element_text}; inherits from \code{title})
 #'                    left-aligned by default\cr
-#'   plot.source      \tab annotations below the plot (text appearance)
+#'   plot.caption     \tab annotations below the plot (text appearance)
 #'                    (\code{element_text}; inherits from \code{title})
 #'                    right-aligned by default\cr
 #'   plot.margin      \tab margin around entire plot