+++ NEWS.md
@@ -7,11 +7,11 @@
 * `ggtitle()` and `labs()` take a `subtitle` parameter which makes it 
   possible to add a subtitle below the main plot title.
   
-* `labs()` takes a `source` parameter that will set the label for 
+* `labs()` takes a `caption` parameter that will set the label for 
   the annotation text below the plot panels.
   
 * The main plot title and subtitle are left-justified (`hjust = 0`) by
-  default. The `source` annotation below the plot will be right-justified
+  default. The `caption` annotation below the plot will be right-justified
   by default.
   
 * Minor code formatting issues in examples and function parameters were